+++ packages/input/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+# @alfalab/core-components-input
 
 ## 15.5.3