+++ packages/input/CHANGELOG.md
@@ -1,5 +1,72 @@
 # Change Log
 
+## 15.5.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.6
+
+## 15.5.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.16.0
+    -   button@11.11.5
+    -   form-control@12.8.2
+
+## 15.5.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.15.0
+    -   button@11.11.4
+    -   form-control@12.8.1
+
+## 15.5.0
+
+### Minor Changes
+
+<sup><time>09.01.2025</time></sup>
+
+### [#1461](https://github.com/core-ds/core-components/pull/1461)
+
+-   Добавлен 40 размер
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.3
+    -   status-badge@1.2.3
+    -   form-control@12.8.0
+
+## 15.4.0
+
+### Minor Changes
+
+<sup><time>26.12.2024</time></sup>
+
+### [#1422](https://github.com/core-ds/core-components/pull/1422)
+
+-   Запрещен ввод и вставка символов `[eE]` в input[type=number]
+
+### Patch Changes
+
+<sup><time>26.12.2024</time></sup>
+
+### [#1502](https://github.com/core-ds/core-components/pull/1502)
+
+-   Апдейт версий пакетов (в них починена сборка esm-версии): @alfalab/data, @alfalab/hooks, @alfalab/utils
+
+-   Обновлены зависимости
+    -   mq@4.4.1
+    -   shared@0.14.1
+    -   status-badge@1.2.2
+    -   button@11.11.2
+    -   form-control@12.7.4
+
 ## 15.3.4
 
 ### Patch Changes